requirements.txt
absl-py==0.9.0
aniso8601==8.0.0
appdirs==1.4.3
asgiref==3.2.10
astunparse==1.6.3
attrs==19.3.0
backcall==0.2.0
bleach==3.1.5
CacheControl==0.12.6
cachetools==4.1.1
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
colorama==0.4.3
contextlib2==0.6.0
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
dill==0.3.2
distlib==0.3.0
distro==1.4.0
Django==3.0.8
djangorestframework==3.11.0
dlib==19.20.0
entrypoints==0.3
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==1.1.2
Flask-RESTful==0.3.8
future==0.18.2
gast==0.3.3
google-auth==1.19.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
googleapis-common-protos==1.52.0
grpcio==1.30.0
h5py==2.10.0
html5lib==1.0.1
idna==2.8
imutils==0.5.3
ipaddr==2.2.0
ipykernel==5.3.2
ipython==7.16.1
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.17.1
Jinja2==2.11.2
joblib==0.16.0
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.6
jupyter-core==4.6.3
jupyterlab==2.2.0
jupyterlab-server==1.2.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
lockfile==0.12.2
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.2.2
mistune==0.8.4
msgpack==0.6.2
mtcnn==0.1.0
nbconvert==5.6.1
nbformat==5.0.7
notebook==6.0.3
numpy==1.19.0
oauthlib==3.1.0
opencv-python==4.3.0.36
opt-einsum==3.2.1
packaging==20.3
pandocfilters==1.4.2
parso==0.7.0
pep517==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
progress==1.5
prometheus-client==0.8.0
promise==2.3
prompt-toolkit==3.0.5
protobuf==3.12.2
psycopg2==2.8.5
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.6.1
pyparsing==2.4.6
pyrsistent==0.16.0
python-dateutil==2.8.1
pytoml==0.1.21
pytz==2020.1
PyYAML==5.3.1
pyzmq==19.0.1
requests==2.22.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.6
scikit-learn==0.23.1
scipy==1.4.1
Send2Trash==1.5.0
six==1.14.0
sklearn==0.0
sqlparse==0.3.1
tensorboard==2.2.2
tensorboard-plugin-wit==1.7.0
tensorflow==2.2.0
tensorflow-addons==0.10.0
tensorflow-datasets==3.2.1
tensorflow-estimator==2.2.0
tensorflow-metadata==0.22.2
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.1.0
tornado==6.0.4
tqdm==4.48.0
traitlets==4.3.3
typeguard==2.9.1
urllib3==1.25.8
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
